+++ assets/js/phoenix_live_view/view.js
@@ -1149,7 +1149,7 @@ export default class View {
         .map(form => {
           // attribute given via JS module needs to be escaped as it contains the symbols []",
           // which result in an invalid css selector otherwise.
-          const phxChangeValue = form.getAttribute(phxChange).replaceAll(/([\[\]"])/g, '\\$1')
+          const phxChangeValue = CSS.escape(form.getAttribute(phxChange))
           let newForm = template.content.querySelector(`form[id="${form.id}"][${phxChange}="${phxChangeValue}"]`)
           if(newForm){
             return [form, newForm, this.targetComponentID(newForm)]
@@ -1167,6 +1167,9 @@ export default class View {
     })
     if(willDestroyCIDs.length > 0){
       this.pruningCIDs.push(...willDestroyCIDs)
+      // we must reset the render change tracking for cids that
+      // could be added back from the server so we don't skip them
+      this.pruningCIDs.forEach(cid => this.rendered.resetRender(cid))
 
       this.pushWithReply(null, "cids_will_destroy", {cids: willDestroyCIDs}, () => {
         // The cids are either back on the page or they will be fully removed,