+++ priv/static/phoenix_live_view.js
@@ -1713,6 +1713,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
                             removeNode(curFromNodeChild, fromEl, true);
                           }
                           curFromNodeChild = matchingFromEl;
+                          curFromNodeKey = getNodeKey(curFromNodeChild);
                         }
                       } else {
                         isCompatible = false;
@@ -1997,8 +1998,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           onBeforeElChildrenUpdated: (fromEl, toEl) => {
             if (fromEl.getAttribute(phxUpdate) === PHX_STREAM) {
               Array.from(toEl.children).forEach((child, idx) => {
-                if (this.streamInserts[child.id].reset) {
-                  this.streamInserts[child.id].streamAt = idx;
+                let insert = this.streamInserts[child.id];
+                if (insert && insert.reset) {
+                  insert.streamAt = idx;
                 }
               });
             }
@@ -2341,6 +2343,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
     getComponent(diff, cid) {
       return diff[COMPONENTS][cid];
     }
+    resetRender(cid) {
+      this.rendered[COMPONENTS][cid].reset = true;
+    }
     mergeDiff(diff) {
       let newc = diff[COMPONENTS];
       let cache = {};
@@ -2519,8 +2524,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       let skip = onlyCids && !onlyCids.has(cid);
       component.newRender = !skip;
       component.magicId = `${this.parentViewId()}-c-${cid}`;
-      let changeTracking = true;
+      let changeTracking = !component.reset;
       let [html, streams] = this.recursiveToString(component, components, onlyCids, changeTracking, attrs);
+      delete component.reset;
       return [html, streams];
     }
   };
@@ -3891,7 +3897,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       let template = document.createElement("template");
       template.innerHTML = html;
       return dom_default.all(this.el, `form[${phxChange}]`).filter((form) => form.id && this.ownsElement(form)).filter((form) => form.elements.length > 0).filter((form) => form.getAttribute(this.binding(PHX_AUTO_RECOVER)) !== "ignore").map((form) => {
-        const phxChangeValue = form.getAttribute(phxChange).replaceAll(/([\[\]"])/g, "\\$1");
+        const phxChangeValue = CSS.escape(form.getAttribute(phxChange));
         let newForm = template.content.querySelector(`form[id="${form.id}"][${phxChange}="${phxChangeValue}"]`);
         if (newForm) {
           return [form, newForm, this.targetComponentID(newForm)];
@@ -3906,6 +3912,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       });
       if (willDestroyCIDs.length > 0) {
         this.pruningCIDs.push(...willDestroyCIDs);
+        this.pruningCIDs.forEach((cid) => this.rendered.resetRender(cid));
         this.pushWithReply(null, "cids_will_destroy", { cids: willDestroyCIDs }, () => {
           this.pruningCIDs = this.pruningCIDs.filter((cid) => willDestroyCIDs.indexOf(cid) !== -1);
           let completelyDestroyCIDs = willDestroyCIDs.filter((cid) => {
@@ -4584,7 +4591,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       return callback ? callback(done) : done;
     }
     pushHistoryPatch(href, linkState, targetEl) {
-      if (!this.isConnected()) {
+      if (!this.isConnected() || !this.main.isMain()) {
         return browser_default.redirect(href);
       }
       this.withPageLoading({ to: href, kind: "patch" }, (done) => {
@@ -4603,7 +4610,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       this.registerNewLocation(window.location);
     }
     historyRedirect(href, linkState, flash) {
-      if (!this.isConnected()) {
+      if (!this.isConnected() || !this.main.isMain()) {
         return browser_default.redirect(href, flash);
       }
       if (/^\/$|^\/[^\/]+.*$/.test(href)) {