+++ test/phoenix_live_view/upload/channel_test.exs
@@ -70,13 +70,16 @@ defmodule Phoenix.LiveView.UploadChannelTest do
   end
 
   def build_entries(count, opts \\ []) do
+    content = String.duplicate("0", 100)
+    size = byte_size(content)
+
     for i <- 1..count do
       Enum.into(opts, %{
         last_modified: 1_594_171_879_000,
         name: "myfile#{i}.jpeg",
         relative_path: "./myfile#{i}.jpeg",
-        content: String.duplicate("0", 100),
-        size: 1_396_009,
+        content: content,
+        size: size,
         type: "image/jpeg"
       })
     end
@@ -100,15 +103,20 @@ defmodule Phoenix.LiveView.UploadChannelTest do
 
   def consume(%LiveView.UploadEntry{} = entry, socket) do
     socket =
-      if entry.done? do
-        name =
-          Phoenix.LiveView.consume_uploaded_entry(socket, entry, fn _ ->
-            {:ok, entry.client_name}
-          end)
+      cond do
+        entry.client_name == "redirect.jpeg" ->
+          Phoenix.LiveView.push_navigate(socket, to: "/redirected")
 
-        Phoenix.Component.update(socket, :consumed, fn consumed -> [name] ++ consumed end)
-      else
-        socket
+        entry.done? ->
+          name =
+            Phoenix.LiveView.consume_uploaded_entry(socket, entry, fn _ ->
+              {:ok, entry.client_name}
+            end)
+
+          Phoenix.Component.update(socket, :consumed, fn consumed -> [name] ++ consumed end)
+
+        true ->
+          socket
       end
 
     {:noreply, socket}
@@ -322,6 +330,17 @@ defmodule Phoenix.LiveView.UploadChannelTest do
           ])
 
         assert render_upload(avatar, "foo.jpeg") =~ "consumed:foo.jpeg"
+      end
+
+      @tag allow: [max_entries: 3, chunk_size: 20, accept: :any, progress: :consume]
+      test "render_upload uploads with progress redirect", %{lv: lv} do
+        avatar =
+          file_input(lv, "form", :avatar, [
+            %{name: "redirect.jpeg", content: String.duplicate("0", 100)}
+          ])
+
+        assert {:error, {:live_redirect, redir}} = render_upload(avatar, "redirect.jpeg")
+        assert redir[:to] == "/redirected"
       end
 
       @tag allow: [max_entries: 3, chunk_size: 20, accept: :any]